# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FertiScan", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/labels": {
      "post": {
        "tags": ["labels"],
        "summary": "Create Label",
        "description": "Create a new label.",
        "operationId": "labels_create_label",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LabelCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelCreated" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["labels"],
        "summary": "Read Labels",
        "description": "List labels with optional filters and sorting.",
        "operationId": "labels_read_labels",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "review_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ReviewStatus" },
                { "type": "null" }
              ],
              "description": "Filter by review status",
              "title": "Review Status"
            },
            "description": "Filter by review status"
          },
          {
            "name": "unlinked",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "description": "Filter labels not linked to a product (product_id is null)",
              "title": "Unlinked"
            },
            "description": "Filter labels not linked to a product (product_id is null)"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Field to sort by",
              "default": "created_at",
              "title": "Order By"
            },
            "description": "Field to sort by"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort direction (asc or desc)",
              "default": "desc",
              "title": "Order"
            },
            "description": "Sort direction (asc or desc)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size limit"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset"
          },
          {
            "name": "product_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Product type",
              "default": "fertilizer",
              "title": "Product Type"
            },
            "description": "Product type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_LabelListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/labels/{label_id}/data": {
      "post": {
        "tags": ["labels"],
        "summary": "Create Label Data",
        "description": "Create LabelData for label.",
        "operationId": "labels_create_label_data",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LabelDataCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelData" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["labels"],
        "summary": "Read Label Data",
        "description": "Get LabelData for label.",
        "operationId": "labels_read_label_data",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelData" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["labels"],
        "summary": "Update Label Data",
        "description": "Update LabelData (partial update).",
        "operationId": "labels_update_label_data",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LabelDataUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelData" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/labels/{label_id}/data/meta": {
      "get": {
        "tags": ["labels"],
        "summary": "Read Label Data Meta",
        "description": "Get LabelDataFieldMeta records for label.",
        "operationId": "labels_read_label_data_meta",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          },
          {
            "name": "field_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Filter by field name",
              "title": "Field Name"
            },
            "description": "Filter by field name"
          },
          {
            "name": "needs_review",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "description": "Filter by needs_review flag",
              "title": "Needs Review"
            },
            "description": "Filter by needs_review flag"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelDataFieldMetaResponse"
                  },
                  "title": "Response Labels Read Label Data Meta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["labels"],
        "summary": "Update Label Data Meta",
        "description": "Update LabelDataFieldMeta (upsert).",
        "operationId": "labels_update_label_data_meta",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelDataFieldMetaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelDataFieldMetaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Contact": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Type of contact",
            "examples": ["manufacturer", "distributor", "importer"]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Company or organization name"
          },
          "address": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address",
            "description": "Full mailing address"
          },
          "phone": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Phone",
            "description": "Phone number in any format"
          },
          "email": {
            "anyOf": [
              { "type": "string", "format": "email" },
              { "type": "null" }
            ],
            "title": "Email",
            "description": "Email address"
          },
          "website": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Website",
            "description": "Website URL"
          }
        },
        "type": "object",
        "required": ["type", "name"],
        "title": "Contact"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelCreate": {
        "properties": {
          "product_type": { "type": "string", "title": "Product Type" },
          "product_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Product Id"
          }
        },
        "type": "object",
        "required": ["product_type"],
        "title": "LabelCreate"
      },
      "LabelCreated": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["id"],
        "title": "LabelCreated",
        "description": "Response schema for label creation."
      },
      "LabelData": {
        "properties": {
          "brand_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name En"
          },
          "brand_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name Fr"
          },
          "product_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name En"
          },
          "product_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name Fr"
          },
          "contacts": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/Contact" },
                "type": "array"
              },
              { "type": "null" }
            ],
            "title": "Contacts"
          },
          "registration_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Registration Number"
          },
          "lot_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Lot Number"
          },
          "net_weight": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Net Weight"
          },
          "volume": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Volume"
          }
        },
        "type": "object",
        "title": "LabelData"
      },
      "LabelDataCreate": {
        "properties": {
          "brand_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name En"
          },
          "brand_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name Fr"
          },
          "product_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name En"
          },
          "product_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name Fr"
          },
          "contacts": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/Contact" },
                "type": "array"
              },
              { "type": "null" }
            ],
            "title": "Contacts"
          },
          "registration_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Registration Number"
          },
          "lot_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Lot Number"
          },
          "net_weight": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Net Weight"
          },
          "volume": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Volume"
          }
        },
        "type": "object",
        "title": "LabelDataCreate"
      },
      "LabelDataFieldMetaResponse": {
        "properties": {
          "field_name": { "type": "string", "title": "Field Name" },
          "needs_review": { "type": "boolean", "title": "Needs Review" },
          "note": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Note"
          },
          "ai_generated": { "type": "boolean", "title": "Ai Generated" }
        },
        "type": "object",
        "required": ["field_name", "needs_review", "note", "ai_generated"],
        "title": "LabelDataFieldMetaResponse"
      },
      "LabelDataFieldMetaUpdate": {
        "properties": {
          "field_name": { "$ref": "#/components/schemas/LabelDataFieldName" },
          "needs_review": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Needs Review"
          },
          "note": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Note"
          },
          "ai_generated": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Ai Generated"
          }
        },
        "type": "object",
        "required": ["field_name"],
        "title": "LabelDataFieldMetaUpdate"
      },
      "LabelDataFieldName": {
        "type": "string",
        "enum": [
          "brand_name_en",
          "brand_name_fr",
          "product_name_en",
          "product_name_fr",
          "contacts",
          "registration_number",
          "lot_number",
          "net_weight",
          "volume"
        ],
        "title": "LabelDataFieldName",
        "description": "Valid field names for LabelDataFieldMeta."
      },
      "LabelDataLite": {
        "properties": {
          "brand_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name En"
          },
          "brand_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name Fr"
          },
          "product_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name En"
          },
          "product_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name Fr"
          }
        },
        "type": "object",
        "title": "LabelDataLite"
      },
      "LabelDataUpdate": {
        "properties": {
          "brand_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name En"
          },
          "brand_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name Fr"
          },
          "product_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name En"
          },
          "product_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name Fr"
          },
          "contacts": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/Contact" },
                "type": "array"
              },
              { "type": "null" }
            ],
            "title": "Contacts"
          },
          "registration_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Registration Number"
          },
          "lot_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Lot Number"
          },
          "net_weight": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Net Weight"
          },
          "volume": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Volume"
          }
        },
        "type": "object",
        "title": "LabelDataUpdate"
      },
...
      "LabelListItem": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "review_status": { "$ref": "#/components/schemas/ReviewStatus" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "label_data": {
            "anyOf": [
              { "$ref": "#/components/schemas/LabelDataLite" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": ["id", "review_status", "created_at", "updated_at"],
        "title": "LabelListItem",
        "description": "Schema for label list items."
      },
...
      "LimitOffsetPage_LabelListItem_": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/LabelListItem" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "minimum": 0.0, "title": "Total" },
          "limit": { "type": "integer", "minimum": 1.0, "title": "Limit" },
          "offset": { "type": "integer", "minimum": 0.0, "title": "Offset" }
        },
        "type": "object",
        "required": ["items", "total", "limit", "offset"],
        "title": "LimitOffsetPage[LabelListItem]"
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": ["not_started", "in_progress", "completed"],
        "title": "ReviewStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}